LayoutTests/fullscreen/video-specified-size-expected.txt

EVENT(webkitfullscreenchange) TEST(video.clientWidth==document.width) OK
END OF TEST
